upating rating done with add, user rating now restricted to 1-5

+++ src/main/java/com/senpro/jafrabackend/models/user/VisitedRestaurant.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Document(collection = "visited")
 public class VisitedRestaurant {
   @Id private VisitedKey id;
-  private float userRating;
+  private double userRating;
   private Date reviewDate;
   @Data
   static public class VisitedKey implements Serializable{
